app/mcp-servers/app-demo-server/requirements.txt
# App MCP Server Dependencies
mcp>=1.0.0
asyncio
typing-extensions>=4.0.0
pyyaml>=6.0
